update support and care url
