+++ src/app/[locale]/support-care/layout.tsx
